9339 Refactoring for hip reads
