+++ app/uk/gov/hmrc/tai/model/domain/Employment.scala
@@ -16,8 +16,7 @@
 
 package uk.gov.hmrc.tai.model.domain
 
-import play.api.libs.json.Reads.localDateReads
-import play.api.libs.json.{Format, JsResult, JsSuccess, JsValue, Json, OFormat, Reads}
+import play.api.libs.json.{Format, Json, OFormat}
 import uk.gov.hmrc.tai.model.domain.income.TaxCodeIncomeStatus
 import uk.gov.hmrc.tai.model.tai.TaxYear
 
@@ -54,50 +53,12 @@ case class Employment(
 
 object Employment {
   implicit val employmentFormat: Format[Employment] = Json.format[Employment]
-
-  val employmentHodReads: Reads[Employment] = new Reads[Employment] {
-    private val dateReadsFromHod: Reads[LocalDate] = localDateReads("dd/MM/yyyy")
-
-    override def reads(json: JsValue): JsResult[Employment] = {
-      val name = (json \ "employerName").as[String]
-      val payrollNumber = (json \ "worksNumber").asOpt[String]
-      val startDate = (json \ "startDate").as[LocalDate](dateReadsFromHod)
-      val endDate = (json \ "endDate").asOpt[LocalDate](dateReadsFromHod)
-      val taxDistrictNumber = numberChecked((json \ "taxDistrictNumber").as[String])
-      val payeNumber = (json \ "payeNumber").as[String]
-      val sequenceNumber = (json \ "sequenceNumber").as[Int]
-      val cessationPay = (json \ "cessationPayThisEmployment").asOpt[BigDecimal]
-      val payrolledBenefit = (json \ "payrolledTaxYear").asOpt[Boolean].getOrElse(false) || (json \ "payrolledTaxYear1")
-        .asOpt[Boolean]
-        .getOrElse(false)
-      val receivingOccupationalPension = (json \ "receivingOccupationalPension").as[Boolean]
-      val status = TaxCodeIncomeStatus.employmentStatusFromNps(json)
-      JsSuccess(
-        Employment(
-          name,
-          status,
-          payrollNumber,
-          startDate,
-          endDate,
-          Nil,
-          taxDistrictNumber,
-          payeNumber,
-          sequenceNumber,
-          cessationPay,
-          payrolledBenefit,
-          receivingOccupationalPension
-        )
-      )
-    }
-  }
-
   private val numericWithLeadingZeros: Regex = """^([0]+)([1-9][0-9]*)""".r
   def numberChecked(stringVal: String): String =
     stringVal match {
       case numericWithLeadingZeros(_, numeric) => numeric
       case _                                   => stringVal
     }
-
 }
 
 case class AddEmployment(
