+++ test/uk/gov/hmrc/tai/model/domain/AnnualAccountSpec.scala
@@ -79,7 +79,7 @@ class AnnualAccountSpec extends PlaySpec {
   )
 
   private def getJson(fileName: String): JsValue = {
-    val jsonFilePath = "test/resources/data/EmploymentHodFormattersTesting/" + fileName + ".json"
+    val jsonFilePath = "test/resources/data/EmploymentHodFormattersTestingNps/" + fileName + ".json"
     val file: File = new File(jsonFilePath)
     val source: BufferedSource = scala.io.Source.fromFile(file)
     val jsVal = Json.parse(source.mkString(""))
